All Codes/storage_to_firestore/requirements.txt
google-cloud-storage
google-cloud-firestore
cloudevents
